fix(ci): drop redundant `as f64` casts; add npm `prepublishOnly` hook
CI clippy (a newer toolchain than my local one) flagged
`clippy::unnecessary_cast` on two factory parsers where `read_number*`
already returns f64. The cast was a leftover from an earlier signature.

Tag v0.2.0 still builds clean — the release workflow only runs
`cargo build`, not clippy — so this lands on main as a follow-up
to keep CI green going forward.

Also wire a `prepublishOnly` script into `crates/ezu-wasm/npm/package.json`
so a manual `npm publish` from that dir transparently rebuilds the
wasm-bindgen outputs first. The CI flow (`./scripts/build-wasm-npm.sh`
+ `npm publish` as separate steps) is unaffected.

+++ crates/ezu-paint/src/nodes/geometry/resample.rs
@@ -131,7 +131,7 @@ impl NodeFactory for SmoothFactory {
         ctx: &FactoryCtx<'_>,
     ) -> Result<BuiltNode, FactoryError> {
         let features = take_input_ref(fields, "features")?;
-        let iter = (read_number_or(fields, "iterations", ctx, 2.0)? as f64).round();
+        let iter = read_number_or(fields, "iterations", ctx, 2.0)?.round();
         if !(0.0..=8.0).contains(&iter) {
             return Err(FactoryError::BadField {
                 field: "iterations".into(),
